added new functions

+++ aliases.zsh
@@ -1,22 +1,28 @@
 # Update Packages
-alias search="brew search"
-alias install="brew install"
-alias update="brew update; sleep 2; brew upgrade; sleep 2; brew cleanup"
-alias uninstall="brew uninstall "
-alias list="brew list"
+if command -v brew &>/dev/null; then
+    alias search="brew search"
+    alias install="brew install"
+    alias update="brew update && brew upgrade && brew cleanup"
+    alias uninstall="brew uninstall"
+    alias list="brew list"
+fi
 
-alias npmsearch="npm search "
-alias npminstall="npm install "
-alias npmupdate="sudo npm upgrade -g"
-alias npmuninstall="npm uninstall "
+if command -v npm &>/dev/null; then
+    alias npmsearch="npm search"
+    alias npminstall="npm install"
+    alias npmupdate="sudo npm upgrade -g"
+    alias npmuninstall="npm uninstall"
+fi
 
-alias pipsearch="pip search "
-alias pipinstall="pip install "
-alias pipupdate="pip list --format=freeze | cut -d = -f 1 | xargs pip install --upgrade"
-alias pipuninstall="pip uninstall "
-alias pipimp="pip install ipykernel matplotlib numpy pandas scikit-learn scipy seaborn torch torchvision torchaudio virtualenv"
-alias pipremove="pip freeze | xargs pip uninstall -y"
-alias pipfreeze="pip freeze > requirements.txt"
+if command -v pip &>/dev/null; then
+    alias pipsearch="pip search"
+    alias pipinstall="pip install"
+    alias pipupdate="pip list --format=freeze | cut -d = -f 1 | xargs pip install --upgrade"
+    alias pipuninstall="pip uninstall"
+    alias pipimp="pip install ipykernel matplotlib numpy pandas scikit-learn scipy seaborn torch torchvision torchaudio virtualenv"
+    alias pipremove="pip freeze | xargs pip uninstall -y"
+    alias pipfreeze="pip freeze > requirements.txt"
+fi
 
 # System
 alias home="cd ~"
@@ -24,7 +30,7 @@ alias ..="cd .."
 alias ...="cd ../.."
 alias c="clear"
 alias e="exit"
-alias dockreset="defaults write com.apple.dock ResetLaunchPad -bool true; killall Dock"
+alias dockreset="defaults write com.apple.dock ResetLaunchPad -bool true && killall Dock"
 alias ll="ls -alF"
 alias la="ls -A"
 alias l="ls -CF"
@@ -32,18 +38,23 @@ alias ls1="ls -1"
 alias mv="mv -v"
 alias cp="cp -rv"
 alias rm="rm -Irv"
-alias sizeof='ls -lh'
-alias f="fuck"
+alias sizeof="ls -lh"
+
+if command -v fuck &>/dev/null; then
+    alias f="fuck"
+fi
 
 # Git
-alias giti="git init"
-alias gits="git status"
-alias gita="git add "
-alias gitc="git commit -m "
-alias gitpull="git pull --rebase origin "
-alias gitpush="git push -u origin "
-alias gitcl="git clone "
-alias gitremote="git remote add origin "
+if command -v git &>/dev/null; then
+    alias giti="git init"
+    alias gits="git status"
+    alias gita="git add"
+    alias gitc="git commit -m"
+    alias gitpull="git pull --rebase origin"
+    alias gitpush="git push -u origin"
+    alias gitcl="git clone"
+    alias gitremote="git remote add origin"
+fi
 
 # Web
 alias brave="open -a 'Brave Browser' --args"
